move bad encoding fraud proof to share pkg (move files)
