+++ share/eds/byzantine/pb/share.proto
